requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
numpy==1.26.4
paddleocr==2.7.3
paddlepaddle==2.6.2
python-multipart==0.0.6
pillow==10.2.0
httpx>=0.25.0
requests>=2.31.0
opencv-python-headless>=4.8.0
